application_src/multi-agent/agent-supervisor/requirements.txt
# Core agent framework
git+https://github.com/strands-agents/sdk-python.git
strands-agents-tools
git+https://github.com/strands-agents/sdk-python.git#egg=strands-agents[a2a]
strands-agents-tools[a2a_client]

# High-performance HTTP client
httpx[http2]>=0.27.0

# High-performance async runtime
uvloop>=0.19.0

# System monitoring for high-RPS scenarios
psutil>=5.9.0

# Advanced caching for better performance
cachetools>=5.3.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Standard dependencies
pydantic>=2.0.0

# Observability dependencies
langfuse==2.50.2
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http

# AWS Dependencies
boto3

# OAuth and JWT authentication - aligned with Configuration API
# Using PyJWT instead of python-jose to avoid CVE-2024-33663 and CVE-2024-33664
PyJWT[crypto]==2.10.1
authlib>=1.3.0
cryptography>=43.0.0

# Security: Pin vulnerable packages to secure versions
requests>=2.32.0  # Fix for GHSA-9hjg-9r4m-mvj7, GHSA-9wx4-h78v-vm56
python-multipart>=0.0.10  # Fix for GHSA-59g5-xgcq-4qw3, GHSA-2jv5-9r88-3w3p
